core/services/bridge/pyproject.toml
[project]
name = "agent-bridge"
version = "0.1.0"
description = "Telegram + Slack messaging bridge for AOS"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "slack-bolt>=1.20.0",
    "slack-sdk>=3.30.0",
    "mistune>=3.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "faster-whisper>=1.0.0",
    "telegraph>=2.2.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "claude_cli", "telegram_channel", "telegram_formatter", "slack_channel", "heartbeat", "activity_client", "voice_transcriber", "interactive_buttons", "daily_briefing", "vault_tasks"]
